chore: add codex migration tooling, patches, and docs
- .gitignore: add .codex-logs/ exclusion, update submodule comment
- scripts/sync-codex.sh: reusable upstream sync with dry-run/check modes
- patches/codex/: build.rs overlay patches for app-server, exec, tui
- docs/maintenance/: migration decision records
- Minor fixes to cargo_workspace_slice.py and language_parser.rs

+++ crates/roo/roo-tree-sitter/src/language_parser.rs
@@ -330,7 +330,7 @@ pub fn process_captures(
                 let child_count = parent.child_count();
                 if child_count > 0
                     && let Ok(last_child_index) = u32::try_from(child_count - 1)
-                    && let Some(last_child) = parent.child(last_child_index)
+                    && let Some(last_child) = parent.child(last_child_index as usize)
                 {
                     let context_end = last_child.end_position().row;
                     let context_span = context_end - parent.start_position().row + 1;
